CI: add job that builds for non-default Python interpreter on Ubuntu
This guards against issues with trying to use the `python3` interpreter
for anything that actually requires the Python interpreter we are
building for. In such a build config with multiple interpreters,
we need to be careful not to pick up the wrong one when invoking
`custom_target` or `run_command` in meson.build files.