TST: ensure `test_extending` compile tests work with any `meson`
The `meson` executable isn't guaranteed to be installed for the same
Python interpreter as we're running the tests for (this may for example
happen in distro setups when we're testing a non-default Python).
Explicitly passing the Python interpreter we're testing for to Meson
via a native file should be a robust solution.