MAINT: numpy cleanup version bumps: fixes issue #20458
After checking discussion to PR #20459, found that the pytest guards could be removed.
So I did, and removed a couple imports and a function which would became unused too.
Also, this commit fixes an horrible error with the previous one, which would ban import prior to NumPy 2.0.0.
(sorry! and thanks @tylerjereddy)
To fix this I change the numpy minimum version of `scipy/__init__.py` to `1.23.5` as user JacobOgle did with PR #20459.
Thanks Jacob!