BUG: interpolate: make BSpline.integrate always return an array
Previously, the return value was either an array (possibly 0D) or
a python float. Given that __call__ always returns 0D arrays, and
so is PPoly.__call__ and PPoly.integrate, there is really no
fundamental reason to not make the return type stable.

Strictly speaking, this is a backwards compat break.