BUG: smoke-docs: ask pytest to not rewrite assertions for doctests
There are no assertions to rewrite in scipy-doctest, and the output
on error is fixed anyway.

This patch follows the suggestion at the very bottom of
https://docs.pytest.org/en/7.1.x/how-to/assert.html

Unless explicitly asked to exclude scipy-doctest from assertion
rewriting, pytest issues an obscure warning on reimport.
The warning is converted into a hard failure in dev.py.