ENH: Updated signal.lombscargle to the generalized Lomb-Scargle periodogram.
The function is a backwards-compatible replacement, but with additional functionality for sample weights and correctly removing any unknown y offset (including where the peaks are clipped, such that a naive mean removal will not work). Additional tests were created for the new options and outputs. Finally, files (and references to them) that are no longer needed were removed (such as the Cython and pythan version of the old code). The compiled versions are not needed, as the vectorized NumPy code is nearly the same speed as the nested for loops.