DOC: optimize: remove inadvertent block quote indentation (#21087)
* DOC: optimize: remove inadvertent block quote indentation

Many lists in the optimize documentation are indented relative to the
surrounding text. The indentation indicates to Sphinx that the
text is to be formatted as a block quote, and our theme represents
that with a vertical bar. Much of this appears to be unintentional.
This commit removes the extra indentation to eliminate these
vertical bars.