BUG: cluster: Avoid OOB write when distances are NaN in `centroid` (#21120)
* BUG: Avoid OOB write when distances are NaN

If a distance matrix with negative distances is passed, to centroid or
median linkage, this can result in NaN distances written back to
the distance matrix. If this happens, none of those distances will be
considered less than inf, so find_min_distance() will return -1, which
will eventually be used to index into an array. This causes an OOB write
and a crash. Forbid this and throw an exception if this ever happens.

* TST: Run negative distance test on cpu backends

---------

Co-authored-by: Jake Bowhay <60778417+j-bowhay@users.noreply.github.com>