MAINT: signal: Don't redefine PyArray_MIN macro.
When building scipy with numpy 1.26.4, the following compiler
warning is generated:

In file included from ../scipy/signal/_splinemodule.cc:5:
../scipy/signal/_splinemodule.h:13: warning: "PyArray_MIN" redefined
   13 | #define PyArray_MIN(a, b) (((a) < (b)) ? (a) : (b))
      |
In file included from ../../../../../py3.12.4/lib/python3.12/site-packages/numpy/core/include/numpy/ndarrayobject.h:12,
                 from ../../../../../py3.12.4/lib/python3.12/site-packages/numpy/core/include/numpy/arrayobject.h:5,
                 from ../scipy/signal/_splinemodule.cc:2:
../../../../../py3.12.4/lib/python3.12/site-packages/numpy/core/include/numpy/ndarraytypes.h:964: note: this is the location of the previous definition
  964 | #define PyArray_MIN(a,b) (((a)<(b))?(a):(b))

In numpy 2.0.0, PyArray_MIN is defined in numpy/npy_math.h, so
it doesn't result in the redefinition problem. In numpy 1.26.4,
PyArray_MIN is defined in numpy/ndarraytypes.h, which is
transitively included by the include of numpy/arrayobject.h
in _splinemodule.cc, resulting in the redefinition warning.

To fix this, the PyArray_MIN macro is replaced with std::min.