BUG: signal: fix crash under free-threaded CPython in medfilt2d
This also improves the logic; the `setjmp`/`longjmp` nor the
`check_malloc` function were good ways to deal with error
handling when memory allocation failed.
It looks like that was introduced because the type-generic
`*_medfilt2d` defined through a macro cannot have an `int`
return type (only `void` for macros)`. However, an integer
error value can instead be passed to the function as an extra
argument.

Closes gh-21142