BLD: remove optional test dependencies from cibuildwheel config [wheel build]
This reverts the addition of gmpy2 from PR 21123, which broke wheels builds
due to missing free-threaded and aarch64 wheels. It also adjusts the comment
above the cibuildwheel test command to explain that we don't want these
optional dependencies but only the essential ones. The full list of test
dependencies is captured higher up in the `test` table under
`[project.optional-dependencies]`.

The goal for wheel builds is reliability, not ultra-comprehensive testing.
That's what we do for optional dependencies in one or a handful of CI jobs.