MAINT: sparse: mark extension modules as safe to run without the GIL
This covers everything except ARPACK/PROPACK, which cannot be done
because the extension module is generated with `f2py` and that doesn't
have a CLI falg to mark a generated module as safe yet. More importantly,
ARPACK and PROPACK are expected to be problematic from a thread-safety
perspective. They're on the list to be rewritten in a language other
than Fortran, so we'll leave them alone here.