ENH: ndimage: add the array API delegation layer
The internal ndimage implementation remains numpy-only, and is intact.

- On CuPy, delegate ndimage.func to cupyx.scipy.ndimage.func
- On other backends, convert inpit arguments to numpy arrays,
  run the numpy-only internal function, and convert the results back
- On JAX, delegate for map-coordinates and convert for the rest.

This way, each ndimage function gets its _dispatcher function, which
knows which arguments are array-like, and deduces the array namespace
based on inputs. This deduction stage needs to be function-specific.
The rest is regular, and is done in _support_alternative_backends.py

Constructing the public API goes through an indirection layer:

- _ndimage_api.py collect imports from private implementation modules;
- _support_alternative_backends.py handles decorates implementor functions;
- __init__.py import from _support_alternative_backends.py