BUG: find_simplex shape () segfault
* Fixes gh-21286

* `find_simplex` has Cython bounds checking disabled, which
means that when it receives an argument for `xi` that `asanyarray`
converts to a 0d NumPy array with shape `()`, the `shape[-1]` check
segfaults before we can even check that the array has the correct
dimensionality for the "point in simplex" algorithm that follows.

* My original reproducer in the ticket above is quite strange--it just
so happens that a `Delaunay` object fed through `asanyarray` has
shape `()`. While this makes basically no sense to me at all, this is
why feeding the triangulation object itself to `find_simplex` behaves
the same as feeding an integer value like `1`, and both segfaulted in
the original ticket.

* Add a regression test for both cases described above/in the ticket,
requiring that an error is raised instead of segfaulting. The tests are
made to pass by using a context manager to locally re-enable Cython
bounds checking for the input dimensionality check.

* Confusingly, the new error raised by Cython when `ndim` is zero
is an `IndexError` (by default), while our custom error checking raises
`ValueError` when `ndim` is too large relative to the generators
used to produce the original triangulation for example. I might be
convinced that we should unify these, though it would require a
try/except to accomplish I believe. For now, I've just added an
additional test for the original `ValueError` (there wasn't one
prior to this patch).

[skip circle] [skip cirrus]