BLD: switch to more robust git hash determination
Fixes the spurious line:

    fatal: bad revision '^v1.13.0'

that shows up in CI logs and locally when the current repo doesn't have a
local v1.13.0 tag.

There is no longer a need to determine a commit count. The method in this commit
(taken over from NumPy) adds the date of the last commit instead. That achieves
the same purpose, namely to have well-defined ordering between two dev builds,
in a simpler way.

This also syncs the rest of version string handling with how NumPy does it;
that NumPy version was done more recently and is more complete.

Note that this also drops the `commit_count` attribute. This field is useless,
and a code search shows it isn't used at all - that seems better than either
making up a wrong number or setting it to None.

That same code search showed that some other attributes are used, so add a test
for them (noting they're not actually public).

Closes gh-21345