DOC: Prolate spheroidal docs wrong
* Fixes gh-20991

* Looking at the original Fortran code:
https://people.math.sc.edu/Burkardt/f_src/special_functions/special_functions.f90
it is clear that the "argument of the radial function" (`x`)
should be strictly `> 1`. See subroutine `rswfp` over there.
Furthermore, we enforce this in the C translation of the code
in our codebase--see the `prolate_radial1` and `prolate_radial2`
function error checking in `sphd_wave.h`.

* While in the neighborhood, fix up other docstring problems related
to these functions, like copy-pase of characteristic value for
functions that don't accept it.

[docs only]