MAINT: Check for all submodule paths in check_installation
As the comment said, the list was getting too annoying, particularly
with the number of tests that would have to be ignored due to the prima
submodule having pybind11 as a submodule (being added in #20964).

It seemed like the ones we want to ignore are the ones in submodules,
since SciPy has no interest in running the tests of its submodules, so I
just implemented some code to find and ignore all submodule paths. At
first it seemed like there might be a bit of trickery with relative
paths, but since CUR_DIR starts as an absolute path, the glob function
returns absolute paths and the `submodule_path in path` check works as
one would expect.