TST: interpolate: fix a spurious failure with -b all
The failing test is testing np.matrix inputs to RGI. This is
allowed for backwards compatibility but does not work with
xp_assert_* infrastructure. Thus, convert to arrays for assertions.

An alternative is to skip the test when running with SCIPY_ARRAY_API
env variable (since then we're not bound by backwards compat). But
converting to ndarray seems simpler.