STY: ignore rule UP038
Rule UP038 recommends using isinstance(a | b) instead of
isinstance((a, b)). This has caused unexpected lint
failures in many CI jobs when files with the older style
are modified, whether or not the offending lines are edited.
I expect this will continue for some time unless either the
rule is ignored or all offending files are fixed at once.
This PR proposes ignoring the rule rather than changing the
codebase, the rationale being that the new style is not
necessarily clearer or otherwise better.

[lint only]