ENH: optimize.least_squares: Add user-defined output function `outfun`
Add an optional argument `outfun` to optimize.least_squares. If `outfun` is
a Callable, it will get called by the optimization algorithm on each iteration.
If the Callable returns True, this causes the algorithm to stop at the current
iteration with status code -2. If it returns False, optimization proceeds as usual.
This can be used to plot the intermediate optimization results or to implement
a user-defined stop condition. By default `outfun` is None, which provides
backward compatibility (if no `outfun` is passed, least_squares behavior remains
unchanged)