BUG: array API conformance in cubature
* Fixes #21745

* The `self._num_inf` object in `_InfiniteLimitsTransform`
results from an array API-based `sum` reduction operation,
which correctly results in a 0d array. `self._num_inf`
is later used in a `reshape` tuple and CuPy is likely
justified in complaining about this because the array
API standard species the `shape` provided to `reshape`
as `Tuple[int, ...]`.

* This fixes the testing issue locally by converting
the 0d array to a Python object using the array API
conformant `__int__()`. Note, however, that this
will/should raise an error for lazy backends. That said,
it seems that local testing with `jax` is "ok," probably
because of skips already present for other reasons.

[skip cirrus] [skip circle]