BUG: special: Fix behavior of `gamma` and `gammasgn` at poles of the Gamma function (#21827)
Fixes the behavior of `special.gamma` and `special.gammasgn` at poles of the `gamma`
function. As described in #21810, we previously had `special.gamma(n) == np.inf` for
every non-positive integer `n`, and bizarrely `special.gamma(-np.inf) == -np.inf`.
Similarly, as described in #19710, we've had `gammasgn(n) == 0.0` for every
non-positive integer `n` and even more bizarrely, `gammasgn(-np.inf) == 1.0`.

The sign of $\Gamma(x)$ at each pole depends on the direction in which $x$ approaches
the pole, so it doesn't make sense to return $+\infty$ at all poles. Accordingly, the
C99 standard recommends the following behavior for the gamma function in special cases.

1. If the argument is $\pm 0$, $\pm \infty$ is returned.
2. If the argument is a negative integer, NaN is returned.
3. If the argument is $-\infty$, NaN is returned.
4. If the argument is $+\infty$, $+\infty$ is returned
5. If the argument is NaN, NaN is returned.

This updates `special.gamma` to follow this standard. Note that signed zeros offer the
ability to specify the sign of infinity of `gamma(x)` at `x == 0`. This also updates
`gammasgn` to be consistent with the above.

Additionally, improve the accuracy of the `rgamma` implementation, and use that as a
replacement where we relied on the previous behaviour; document this in the docstring.