MAINT: Hausdorff Generator handling
* This patch is a precursor to SPEC 7 compliance of
`directed_hausdorff` requested in gh-21833. I believe
the team prefers that I first add support for `Generator`
handling here, and once they are happy with that they will
follow up with the usual decorator magic.

* In addition to modernization of the source proper to accept a `Generator`,
some test cases using a `Generator` with the "same" integer
seed as the previous `RandomState` approach have been added.
One such case demonstrates that defaulting to `Generator` in the
long term will change some results, which SPEC 7 does acknowledge.
For clarity, the returned Hausdorff distance will always remain identical,
but the indices corresponding to that distance may vary according to
the exact random status for this shuffling-reliant
algorithm. Disruption is likely to be minimal if we follow the SPEC
process/warnings timeline I'd say. Of course, if we follow
the full procedure the indices returned for the same function
call can change for some inputs.

* Note that there are type stub/`pyi` signature(s) that might
in principle be changed here but `mypy` didn't seem to care locally
and IIRC our recent philosophy has been to delegate typing
improvements to an external package.

[skip cirrus] [skip circle]