BUG: fix nogil LinearNDInterpolator
* Fixes gh-21885

* A `nogil` block in `LinearNDInterpolator` is reliably
segfaulting on x86_64 Ubuntu Linux with Python 3.10 and
Python 3.11. I see the same error locally as reported in CI,
and it is present in both `v1.14.1` and on latest `main`
branch of SciPy. Diffing the last succeeding version
of the CI job with the current failing CI job only shows
`tomli` and `hypothesis` changing versions. So, this is pretty
weird, but the test is behind a `slow` marker so maybe we're just
discovering the instability now for some reason.

* Removing the `nogil` directive fixes the issue on Python `3.10`
and `3.11` locally.

* Note that GNU compiler toolchains 9.x and 11.x both reproduce
the segfault locally as well, so the "minimum version" CI label may
be a red herring. It is also worth noting that the CI job name
contains "fast," and the `pytest` incantation isn't using slow,
so there may be yet another issue about i.e., respecting markers
when running `pytest` outside the root folder in that job or something
like that. Nonetheless, the test fails consistently before this fix
locally, which should probably supersede performance concerns for now.

[skip circle] [skip cirrus]