TST: temp skip of extending tests
* Temporarily skip the tests flagged in gh-22039
on free-threaded Windows CPython. Note that free-threading
itself may be a red herring here, and the issue may ultimately
be some combination of the disabling of build isolation
for free-threaded wheels compounded with the Windows platform.

* Note that the error messages reported by these test meson
builds are identical to those reported by in-place builds
of SciPy, for which we already have skips on these tests.
Even for the case of the in-place build test failure,
which simplifies down to a single-line reproducer, `cimport scipy.linalg`,
we still don't have a work around, so it seems like we may
not want to hold up the release candidate for an even more
challenging-to-reproduce scenario that has the same error message.
Also, if there were fundamental issues with free threading
on this platform/Python combination, they would presumably
show up somewhere other than just these in-place-sensitive
tests.

[skip circle] [skip cirrus]