ENH: Implement Muller's method in scipy.optimize._zeros_py.py and _root_scalar.py
Reference issue

Toward #21126

What does this implement/fix?

This is an implementation of Muller's method (as discussed in #21126). It supports vectorized operations. The method is usable from `root_scalar()` as well.

Additional information

I've decided to implement this as a separate method inside of `_zeros_py.py`. We discussed adding it to `newton()` but the function is already quite unwieldy. Given that, and the more specialized nature of this algorithm, I've decided to make it into it's own function. Of course, pending discussion, the function can be made private and simply called from `newton()`.
Usefulness of this function lies in the fact that it has reasonable convergence (slightly higher than the secant method), and unlike Newton's method it doesn't require a derivative. It also converges towards complex roots even with completely real initial guesses. My primary use of this is searching for multiple roots using a `np.linspace` of initial guesses in a certain region.

This is the first time I'm contributing to SciPy, so I'm looking forward to your helpful suggestions. I've tried to mimic the implementation of `newton()` for consistency with existing codebase. There are some notes in the comments as well regarding the specific implementation decisions and why were they chosen over alternatives.