TST: TestBracketMinimum MPS shims
* Two tests in `TestBracketMinimum` were failing with the
MPS backend + `torch` on ARM Mac. The tests were asserting
for exact equality between floats--it seems that if we just
do the usual thing of `xp_assert_close()` when comparing
floats the test passes just fine without needing additional
checks for default dtype, and this seems like the right
approach even if the tests were not failing.

[skip circle] [skip cirrus]