MAINT: Fix dtype check in `scipy.signal._waveforms`.
The old code used to call `dtype.char` and then check if it was
contained with a list of a single string `['fdFD']`. This of course
would never pass because a `'f'` is not a member of `['fdFD']` but it is
a member of `'fdFD'`.

This is likely a mistake in the code and so this commit fixes it.