ENH: io.wavfile: read unseekable files (#22208)
* ENH: read unseekable files in scipy.io.wavfile.

Allow scipy.io.wavfile.read() to read non-seekable files.

  * If passed stream is not seekable, wrap it in a SeekEmulatingReader,
    which tracks current position and implements tell() to report it.

Most seek()s in wavfile.read() simply seek forward from the current
position by a fixed amount. To address these:

  * Support limited SeekEmulatingReader.seek() that emulates seek()s
    that can be supported by read()ing and discarding the result.

Befor returning, read() attempts to seek back to the start of the file.
This is nice when possible, but if it isn't we shouldn't blow up. So:

  * Avoid final fid.seek(0) if underlying stream isn't seekable.

The remaining seek()s are in RF64 support, where chunk size is global
and stored in the RIFF header rather than in the chunk header.
_read_data_chunk() currently seeks back to a fixed offset from the file
origin to grab the chunk size. To address this:

  * Modify RF64 support to parse and store the chunk size in
    _read_riff_chunk() so that _read_data_chunk() doesn't have to seek()
    back for it. This applies whether or not reading from a seekable
    stream.

To test this, I added test_streams(), which ensures that wavfile reads
the same data no matter whether a path, seekable stream, or unseekable
stream is passed to read().

Fixes #11328

* TST: assert_array_equal -> assert_equal

---------

Co-authored-by: Jake Bowhay <60778417+j-bowhay@users.noreply.github.com>