BUG: Add explicit recasts of intermediate arrays in `sawtooth`
Before this commit, the `sawtooth` function in
`scipy.signal._waveforms.py` could error out on numpy v1.x due to a
change in type promotion rules between v1 and v2.

Prior to numpy 2, type promotion was value based meaning the intermediate
values could be promoted from `float32` to `float64` due to use of
python `float` scalars, which then promotes the output dtype to
`float64`, which leads to errors as the `place` function uses safe type
casting.

This patch avoid this by casting the intermediate arrays back to the
output dtype for each call of `place`.