DOC: Update COBYLA docs
It no longer wraps a Fortran implementation. I'm not sure why the line
about the function returning a list of number was there in the first
place. It doesn't make much sense, and also doesn't work as of this
writing with the existing implementation and certainly not with the new
one.