BLD: bump min version of Clang to 15.0, and macOS min version to 10.14
Addresses the wheel build failures reported for macOS x86-64 in gh-22487.

macOS 10.14 is from 2018, so is still a very loose lower bound, we don't expect
issues as a result.

The Clang >=15 is probably also reasonable, since it's rare to see a version
below 16 being used. Targeting macOS 10.13 causes cibuildwheel to pull in
XCode 14.1 with Clang 14.0, while macOS 10.14 pulls in XCode 15.2 and Clang 15.0.