ENH: Adding linewalker method to minimize_scalar
Linewalker constructs a smooth surrogate on a set of equally-spaced
grid points by evaluating the true function at a sparse set of
judiciously chosen grid points. At each iteration, the surrogate's
non-tabu local minima and maxima are identified as candidates for
sampling. Tabu search constructs are also used to promote
diversification. If no non-tabu extrema are identified, a simple
exploration step is taken by sampling the midpoint of the largest
unexplored interval. The algorithm continues until a user-defined
function evaluation limit is reached.

Linewalker is particularly well-suited for nonconvex (multimodal)
functions. If the underlying function is known to to be convex,
use a local solver.