MAINT: Remove an extraneous dtype check in `scipy/signal/_waveforms.py` (#22520)
* MAINT: Remove dead code in waveforms' sawtooth and square

The dead code refers to a dtype check that is wrongly checking if the
dtype char (which is a single character string) is contained within a
list of a single string. This typo has meant that this check can never
pass because a single character string can never be an element of the
given list.

In order to clean this up while maintaining backwards compatibility, I
have removed the if statement and now both functions always return
np.float64 arrays explicitly.