BENCH: remove triple run of mmread/mmwrite benchmark, limit sizes
This fixes a problem on CircleCI. Closes gh-22574

This also reduced the runtime of the `io_mm` benchmarks by 10x,
it was super slow for no good reason because the benchmarks were
run 3 times, 2x with fast-matrix-market and 1x with the old pure
Python implementation. This looked like a left-over from the PR
that switched to fast-matrix-market, demonstrating that `io.mmread`
and `io.mmwrite` are the fast version and `io._mmio` is way slower.
There is no need to keep running that.