BUG: `lfiltic`'s handling of `a[0] != 1` differs from `lfilter`'s  (#22554)
* Added test to highlight issue #22470 prior to fixing it

* BUG: lfiltic's handling of a[0] != 1 differs from lfilter's

Fixes 22470 by dividing the resultin zi by a[0].
Also added checks that a is non-empty and its first element is non-zero

Closes #22470

* Addressed CI error and added test coverage for ValueErrors (as requested by j-bowhay)

* Small changes to address CI errors

* Fixed failing test with cupy and excluded test with jax.
The problem with jax is that does not support inplace ops.
It could be easily fixed, e.g. with zi = zi.at[idx].add(...)
Another option would be to rewrite parts of the code to use
(xp or scipy) convolution instead.

* Fixed Lint error UP038 Use `X | Y` in `isinstance` call instead of `(X, Y)`