ENH: Replace Fortran COBYLA with Python version from PRIMA (#22350)
* DEV: Don't check submodules in check_python_h_fist

* ENH: Replace Fortran COBLYA with Python version from PRIMA

PRIMA now has a pure python implementation of COBYLA!

This commit also
- Updates COBYLA to accept the new style constraints
- Adds several options, including ftarget and more granular iprint
- Updates COBYLA to use actual bounds and linear constraints instead of
  simply adding them to the nonlinear constraint
- Updates COBYLA to use the new style callback interface
- Updates COBYLA to be threadsafe
- Updates all relevant documentation
- Removes the old implementation

* TST: Remove usage of assert_ from test_cobyla

* DOC: Update COBYLA docs

It no longer wraps a Fortran implementation. I'm not sure why the line
about the function returning a list of number was there in the first
place. It doesn't make much sense, and also doesn't work as of this
writing with the existing implementation and certainly not with the new
one.

* DOC: Add release notes for COBYLA

* TST: Give a little more time for the basinhopping test [wheel build]

* MAINT: remove prima submodule

* MAINT: add pyprima submodule

* BLD: update meson for pyprima

* switch to alternative script

* delete lucascolley/pyprima submodule

* add scipy/pyprima submodule

* Remove tach check on PRIMA

PyPRIMA imports scipy in order to use the constraint objects, which
results in a circular dependency violation. COBYQA does the same thing,
so like with COBYQA we set unchecked to true.

---------

Co-authored-by: Lucas Colley <lucas.colley8@gmail.com>
Co-authored-by: Daniel Schmitz <40656107+dschmitz89@users.noreply.github.com>