ENH: spatial: speed up `Rotation.apply` by replacing `np.einsum` with `np.matmul` (#20950)
* ENH: spatial: Rotation.apply is slow - replace np.einsum with np.matmul

Replacing the use of np.einsum in scipy.spatial.Rotation.apply with
np.matmul speeds up the function by more than a factor of 2. The
calling semantics is a bit more complex, covering all cases
of dimensionality of Rotation and vectors.

* ENH: spatial: Rotation.apply is slow - replace np.einsum with np.matmul

Reverts back one of the code path to use np.einsum, as it is faster and adds a test for the shape of the output array.

* ENH: spatial: Rotation.apply is slow - replace np.einsum with np.matmul

By mistake added a space, fixed it.

* ENH: spatial: Rotation.apply is slow - replace np.einsum with np.matmul

Removes unnecessary np.all from test_apply_shape

* breaks up Rotation.apply in a check and compute block

* adds types and shorten checks for rotations