MAINT/TST: rewording of "ties" into "tied pairs" for clearer meaning (#22859)
* Rewording of "ties" into "tied pairs" for clearer meaning

This is a very simple modification that I thought could be useful for people new to this metric.
The use of "ties" in this context is misleading, possibly hinting to only counting tied elements and not actually all the possible tied pairs. This looks more appropriate and also keeps the phrase consistent, together with "concordant pairs" and "discordant pairs".

* lint

---------

Co-authored-by: Lucas Colley <lucas.colley8@gmail.com>