BUG: add a fudge factor for jax in findfreqs
Otherwise, xp.round(-1.5) is -2 or -1 depending on fp jitter.
While this is a bad form, tests only catch the issue with jax.numpy
on either GPU or CPU, and not any other backend.

And the computation in findfreqs is prone to this kind of instability, yes.
Am not going to change it right now because backwards compat though.