ENH: enable passing parameters to differential evolution polishing step
As a final step, differential evolution calls the minimize function to "polish" the results. It would be useful to be able to pass additional parameters to the minimize function (e.g. a jacobian or the name of a minimization method). The dual annealing method already offers this option. This commit adds this possibility for differential evolution via the minimizer_kwargs parameter. The default behavior is unchanged.