TST: ndimage: avoid implicit dtype conversions in vectorized_filter tests
This tests does this:  `vectorized_filter(input, xp.sum, output)`, with output
being `xp.empty_like(input)`. This is problematic for short integers because
- the output of xp.sum(int8_array) is int64 (the default int dtype), and
- internally, vectorized_filter does roughly,
`output[some_indices] = xp.sum(input[some_indices])`

So what happens here is that the output of the `sum` in the r.h.s. gets silently
downcast to the dtype of the l.h.s. by `__setitem__`.

This kind of sort of works in many array libraries, but is technically unspecified
by the Array API standard. And indeed this starts failing with array-api-strict
in https://github.com/data-apis/array-api-strict/pull/157

Thus, a test only change to make downcasting explicit.