ENH: Compute exact 1-norm of A in cond1est.
Needs performance testing, but the condition number estimate will be
more accurate if we are not multiplying two estimates together. The
sparse 1-norm computation of the original matrix should be efficient.

Note that the __init__.py imports needed to be reordered so that
scipy.linalg.norm is available in cond1est.