TST: Refactor SEED to rng instance.
This commit decouples any test fixtures called at the same time with the
same seed. It revealed a bug in generate_matrix where a given "exactly
singular" matrix may raise a different error from splu, so this commit
also updates the construction of the random matrices to ensure we get an
exactly singular matrix when requested.