BUG: interpolate.RegularGridInterpolator: fix linear interp. on 2D grid with a single point (#23253)
* TST: interpolate.RegularGridInterpolator: test on single-point grid

(reviewed at https://github.com/scipy/scipy/pull/23253) 

New method `TestRegularGridInterpolator.test_length_one_axes` checks
that `interpolate.RegularGridInterpolator` works as expected for
methods 'linear' and 'nearest' if the grid contains a single point
(in all dimensions).

As of 4c1a4da81 this fails for the following parametrization
`TestRegularGridInterpolator::test_length_one_axes[linear-2-float64]`,
but is OK in all other cases.

See GH-23171.

* BUG:interpolate.RegularGridInterpolator: fix linear interp. on 2D grid with a single point

Explicitly handle the case in which there is a single grid point on both
axes.

Branches in which `out[point] = NAN` is set are deleted because
redundant: out-of-bounds points/nans are handled by caller.

* Update scipy/interpolate/tests/test_rgi.py

Co-authored-by: Evgeni Burovski <evgeny.burovskiy@gmail.com>

* MAINT: delete asserts from code

* TST: refactor test method

- Rename `test_length_one_axes` as `test_length_one_axis_all`.
- Split tests for various types of input points into separate asserts.
- Use more idiomatic array API functions.

* TST: delete spurious `assert_array_equal`

Make clear that it is the interpolator call that fails,
`not assert_array_equal`.
Use a slightly more verbose call pattern: best practice is to have
a bare assert, without inside calls.

* TST: Implement xp_assert_equal

- explicitly set wanted dtype
- cleanup minor code nits

---------

Co-authored-by: Evgeni Burovski <evgeny.burovskiy@gmail.com>