ENH: Compute cond1est as SuperLU member function. Remove normest_inv.
This commit addressed reviewer comments on PR #23292 regarding internal
computation of `cond1est` vs. computing two norms in Python.

This commit makes an internal change to the SuperLU object to compute
and store the 1-norm of the matrix A when splu(A) is called. That value
is then used by the new member function `SuperLU.cond1est`, which calls
`gscon` to compute the condition number.

This commit removes the existing `SuperLU.normest_inv` member function,
which is no longer necessary. It also removes the associated test class
TestNormEstInv.