BENCH, TST: benchmarks without hypothesis
* Fixes gh-23344.

* The new array API `asv` benchmarks depend on `scipy.conftest`,
which has a mandatory dependency on `hypothesis`. This patch
makes it possible to extract the necessary array API information
from `conftest` without needing the `hypothesis` dependency, so
that benchmarks can run in that situation.

* I'm not sure if it is desirable long-term to have this kind
of cross-talk between the `asv` benchmarks and our testing
infrastructure, but this seemed simplest for now. Note that
`hypothesis` is still mandatory for running our full testsuite.