BLD: add Accelerate ILP64 support
This is a follow-up to scipy#22743, fixing a few small issues with that
PR. That makes Accelerate build and test fine on macOS >=13.3 with
the latest Meson release, the only manual bit being that one has to
add the BLAS symbol suffix manually with `-Dblas-symbol-suffix=$NEWLAPACK$ILP64`.

In a Pixi dev setup, this task works:
```
[feature.accelerate.tasks.build-accelerate-ilp64]
cmd = "spin build --build-dir=build-accelerate-ilp64 --with-accelerate -S-Duse-ilp64=true -S-Dblas-symbol-suffix='$NEWLAPACK$ILP64'"
```

The one hiccup is an interpolate test for integer overflow that starts crashing,
see issue scipy#23396. Skipping that problematic test for now.